move screen prop interfaces to own file
