+++ thoughtbubble-mobile/src/interfaces/componentProps.ts
@@ -1,10 +1,5 @@
-import { RouteProp } from '@react-navigation/native';
-import { StackNavigationProp } from '@react-navigation/stack';
 import { ProjectShape, ThoughtShape } from './data';
-import { ProjectStackParamList, StatsStackParamList } from './navigation';
 
-export interface LoginScreenProps {}
-export interface SettingsScreenProps {}
 export interface AppNavTabsProps {}
 export interface ProjectsStackNavigatorProps {}
 
@@ -77,27 +72,4 @@ export interface ThoughtsListProps {
   handleThoughtStatusChange: (thoughtId: string) => void;
   handleThoughtDelete: (thoughtId: string) => void;
   handleScroll: (...args: any[]) => void;
-}
-
-// ===== screen props =====
-
-// for central projects/thoughts nav stack
-// typing reactnavigation -> https://reactnavigation.org/docs/typescript/
-export interface ProjectsScreenProps {
-  navigation: StackNavigationProp<ProjectStackParamList, 'Projects'>;
-}
-
-export interface ThoughtScreenProps {
-  route: RouteProp<ProjectStackParamList, 'Thoughts'>;
-  navigation: StackNavigationProp<ProjectStackParamList, 'Projects'>;
-}
-
-// for stats nav stack
-export interface StatsHomeScreenProps {
-  navigation: StackNavigationProp<StatsStackParamList, 'Analytics'>;
-}
-
-export interface StatsProjectInfoScreenProps {
-  route: RouteProp<StatsStackParamList, 'Project Analytics'>;
-  navigation: StackNavigationProp<StatsStackParamList, 'Analytics'>;
 }
